LayoutTests/inspector/styles/stylesheet-tracking.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<link rel="stylesheet" href="resources/stylesheet-tracking.css" />

<style>
body {
    font-size: 12px;
}
</style>

<script>

function addStyleSheet()
{
    var styleElement = document.createElement("style");
    styleElement.id = "style";
    document.head.appendChild(styleElement);
}

function removeStyleSheet()
{
    document.head.removeChild(document.getElementById("style"));
}

function test()
{
    var inspectorResource;

    WebInspector.showPanel("elements");
    WebInspector.cssModel.addEventListener(WebInspector.CSSStyleModel.Events.StyleSheetAdded, styleSheetAdded, null);
    WebInspector.cssModel.addEventListener(WebInspector.CSSStyleModel.Events.StyleSheetRemoved, styleSheetRemoved, null);
    var headers = WebInspector.cssModel.styleSheetHeaders();
    InspectorTest.addResult(headers.length + " headers known:");
    for (var i = 0; i < headers.length; ++i)
        InspectorTest.addResult(i + ":\n" + headerData(headers[i]));

    step1();

    function step1()
    {
        InspectorTest.addResult("Adding stylesheet...");
        InspectorTest.evaluateInPage("addStyleSheet()", step2);
    }

    function step2()
    {
        InspectorTest.addResult("Removing stylesheet...");
        InspectorTest.evaluateInPage("removeStyleSheet()", step3);
    }

    function step3()
    {
        InspectorTest.selectNodeAndWaitForStyles("inspected", step4);
    }

    function step4(node)
    {
        InspectorTest.addResult("Adding rule...");
        WebInspector.cssModel.addRule(node.id, "#inspected", successCallback, failureCallback);

        function successCallback()
        {
            InspectorTest.addResult("Rule added");
            InspectorTest.completeTest();
        }
        function failureCallback()
        {
            InspectorTest.addResult("Failed to add rule.");
            InspectorTest.completeTest();
        }
    }

    function styleSheetAdded(event)
    {
        var header = event.data;
        InspectorTest.addResult("Stylesheet added:\n" + headerData(header));
    }

    function styleSheetRemoved(event)
    {
        var header = event.data;
        InspectorTest.addResult("Stylesheet removed:\n" + headerData(header));
    }

    function headerData(header)
    {
        return "    URL=" + trimURL(header.sourceURL) + "\n    origin=" + header.origin;
    }

    function trimURL(url)
    {
        if (!url)
            return url;
        var lastIndex = url.lastIndexOf("inspector/");
        if (lastIndex < 0)
            return url;
        return ".../" + url.substr(lastIndex);
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that the styleSheetAdded and styleSheetRemoved events are reported into the frontend. <a href="https://bugs.webkit.org/show_bug.cgi?id=105828">Bug 105828</a>.
</p>

<div id="inspected">Text</div>

</body>
</html>
